Add Razor shims
Add Razor Language Service shim for accessing the remote client

Add Razor Remote Workspace shim for getting access to the solution in
the ServiceHub process. This relies on the same set of Roslyn services
being called by the client to initialize state in the remote process.

Add overloads to the 'remote' solution service to avoid a need for IVT in
in the Razor Remote Workspace shim. Checksum is internal to Workspace.