Don't assert if an option wasn't initialized that never will be
The TryFetch call should never get called for options we'll initialize
during the constructor, which is what this assert was intended to catch.
But we'd still fire if we saw a language we didn't know about. Although
this has caught some legitimate bugs, at this point it's not terribly
valuable, and also hurts cases where other languages (like interactive
commands) don't need this in the first place.

Fixes dotnet/roslyn#14976.