Detect failed server connections in the bootstrap build
This change will cause our bootstrap build to fail if a threshold of server connection failures
occur.  The threshold is chosen a bit arbitrarily as valid failures can occur.  But 15 seems like
a number to warrant an investigation.  If this produces too many false positives it will be
adjusted appropriately.