Enable F5 / TDD testing of our portable unit tests
At this moment there is no supported way of building a portable DLL and deploying
current OS dependencies (say windows) as a default.  As a result these DLLs are
not full deployed post build (not runnable) and this breaks any number of
tools developers use to be productive: F5, TDD, WPF xunit runner.

Roslyn works around this for official builds by using deployment projects.  This
is a project per supported delpoyment (Desktop, CoreCLR) which consists of a
dummy EXE and referenes all of our portable DLLs.  Hence they DLLs as they exist
in that output directory are fully deployed.

This makes it possible to do official testing with confidence but doesn't developers
who want a more rapid feedback experience.

This change makes it possible to do direct testing again by adding an ad-hoc deployment
mechanism.  This uses the same code as the official deployments but is still ad-hoc
so it won't be used for official builds: just developer builds.

Here is the bug on MSBuild tracking the real bug and fix

https://github.com/Microsoft/msbuild/issues/1499