Re-arranging portable TestUtilities part 2
The Compilers.sln solution is compiling again and tests which don't
require AppDomain support are running again.  There is a bit of work
that is needed for AppDomains that I want to separate out.  In
particular the separation to portable means we have to redo some of the
serialization code.