Move to Microsoft.Tpl.Dataflow
It would be nice to move to System.Threading.Task.Dataflow however VS componentss we depend on depend on MS.Tpl.Dataflow.  Creates a bit of a packaging nightmare. They won't move to the roslyn alpha version because it's not pubilc.  Hence we need to use the standard one until this bug is fixed.

https://github.com/dotnet/corefx/issues/14869