Allow the C# compiler to run on Mono again
**Customer Scenario**: Referencing the latest Microsoft.Net.Compilers package and try to run csc.exe on Mono fails. (This was
found building the xplat branch of MSBuild.)

**Summary of Change**: Don't use System.Runtime.InteropServices.RuntimeInformation in Compiler, Workspace or Features layers

**Details**: Recently, we took a change to use `System.Runtime.InteropServices.RuntimeInformation.IsOSPlatform(...)` to detect the
current platform. The API is designed such that different implementations are delivered via NuGet depending on the TFM. For .NET
Core, this works great. When restoring an app using Microsoft.CodeAnalysis on Windows, a Windows implementation is delivered. And,
when restoring the same app targeting macOS/Linux, a Unix-specific implementation is delivered. However, for apps targeting
.NET Framework (such as the desktop compiler), this design presents a serious problem. Essentially, the Windows implementation
is always picked for .NET Framework apps (such as net451, net46, etc.). So, if that app is then run on macOS or Linux on Mono,
Microsoft.CodeAnalysis will *think* it's running on Windows and start reporting errors because path separators are pointing the
wrong direction.