Don't defer the call to CreateEventingTaskQueue()
If we try deferring this, we might end up calling it on the first thread
to raise an event, instead of thread that created the workspace, which
was our previous contract. The default implementation looks at whether
there is a SynchronizationContext to determine what the scheduler is
to use, so the change matters.