Avoid deadlocks when unloading .NET Standard Projects (directly or via Solution close) referencing Shared Projects
Fixes https://github.com/dotnet/roslyn/issues/14479

The problem here is that unloading projects takes a serialization lock,
which can cause Shared Project IVsHierarchys to notify us that their
context hierarchy has changed. In response to this, we try to set the
new active context document for open files, which also tries to take the
serialization lock, and we deadlock.

For .NET Framework Projects that reference Shared Projects, two things
prevent deadlocks when projects unload. During solution close, any
Shared Projects are disconnected before the projects start to unload, so
no IVsHierarchy events are fired. During a single project unload, we
receive notification of the context hierarchy change before the project
is unloaded, avoiding any IVsHierarchy events if we tell the shared
hierarchy to set its context hierarchy to what it already is.

Neither of these behaviors are safe to rely on with .NET Standard
projects, so we have to prevent the deadlock ourselves. We do this by
remembering if we're already in the serialization lock due to project
unload, and then not take the lock to update document contexts if so
(but continuing to lock if it's not during a project unload).