Fix dynamic analysis instrumentation for VB
We insert a bound node of method call to `CreatePayload(Guid,int,int,ref bool[],int)` during instrumentation for each method,
however, the node of the 4th argument (of type bool[]), which is an array access (something like payloads[2],
where payload is of type bool[][]), is created as an RValue. As an result, in the IL generated for instrumentation,
the argument is the reference to an local variable instead of to the indexed element in the array. This will cause `CreatePayload`
to throw an exception when it tries to add a payload array for a method that already has a payload array created.