Move `WaitForAsyncOperations` call out of tests
And into `FindReferencesWindow_OutOfProc.GetContents`. This way individual tests don't need to remember to wait for any pending Find References operations to complete.

Also, make use of xUnit's `Assert.Collection(...)` 	method to verify collection contents. Specifically we don't want to use the LINQ `Single()` method--if it throws it won't cause us to take a screenshot, as we only do that for `XunitException`s.