Always take screenshots when an exception occurs
Currently our integration tests only take a screenshot when an `XunitException` occurs. This is too narrow; other kinds of exceptions may occur and cause the test to fail, and we currently get no screenshots for those. This is especially problematic when an exception occurs in the VS process and is remoted and rethrown in the test process.

The only downside I see is that we will state to get screenshots for expected exceptions, but that seems like a very minor price to pay.