Add a helper method for getting a dialog
Right now a couple of methods use `DialogHelpers.FindDialog` to try to locate an open dialog, and then throw an `InvalidOperationException` if they don't find it. This is going to be a pretty common pattern when working with dialogs and controls within dialogs, so I've refactored it into a new `DialogHelpers.GetOpenDialog` method.

I've also updated the `Editor_InProc.PressDialogButton` method to make use of the `AutomationElement` extension methods.