Fix SolutionLoadEvents
So, it turns out that when you AdviseSolutionEvents, the shell actually
checks the same object for all of the related solution events interfaces
to fire events to them. This meant that after commit
e6a43cb, we never actually received any
IVsSolutionLoadEvents or IVsSolutionWorkingFoldersEvents. The impact of
this is that we never called "FinishLoad" to populate the Roslyn
workspace with the projects that had been loaded.

Most features would appear to work, as we would load a project and it's
dependencies when a file was opened, but we would then not load any
thing that depends on it, so features like Find All References and
Rename wouldn't work reliably.

Fixes
https://devdiv.visualstudio.com/DevDiv/_workitems?_a=edit&id=389698.