Have Code DOM tests check for ComDefaultInterface
The Code DOM tests which verify the behavior of
TypeDescriptor.GetProperties relied on certain type libraries being
registered machine wide, EnvDTE in particular.  This is not true when
running our tests on bare Windows machines.

Changed these tests to instead directly validate the ComDefaultInterface
value for the types.  This is what ultimately controls the values
returned by GetProperties and hence is a sufficient test.