Add runtime reference to MS.VS.Debugger.Metadata
These EE projects tranisively reference MS.VS.Debugger.Metadata but it
is not referenced.  Hence it was not a part of the project output and as
a result test libraries did not participate in caching.

Adding this reference directly as compile + runtime created a lot of
compilation errors due to the following:

- The DLL contains a namespace named Microsoft.MetadataReader
- The projects make heavy use of System.Reflection.MetadataReader
- The project namespace is Microsoft.*

Once a compile reference was added to MS.VS.Debugger.Metadata it caused
MetadataReader to prefer the namespace over the type.  This is fixable
by adding a lot of using directives into ~5 source files but that seemed
fairly disruptive hence I went with this approach.