Support diagnostic nodes in the CPS-based project system
When using the legacy project system the "Analyzers" node under the "References" node in Solution Explorer, and all of its descendants, are implemented via the IAttachedCollectionSource interface. This is because the project system does not surface them in the project's IVsHierarchy, and we didn't want to invest the time and effort required for such a change when we were already thinking about moving to a CPS-based project system. IAttachedCollectionSource allowed us to implement this feature in a simpler way.

In the new CPS-based project system, however, analyzers are represented in the IVsHierarchy as they are proper project items. The nodes for individual diagnostics are not part of the project, and so it makes sense to continue to implement them using IAttachedCollectionSource.

This change copies the `DiagnosticItemProvider`, `DiagnosticItemSource`, and `DiagnosticItem` types and modifies them to create equivalents that will work with CPS. The most substantial modifications are made to the `CpsDiagnosticItemSource` type. In the CPS world we may have a Solution Explorer node representing the analyzer before any information about the project or analyzer has been pushed to the workspace, and thus we need to handle the situation where the `CpsDiagnosticItemSource` is created before it can actually provide any items. In this case it listens to `WorkspaceChanged` events and, when the information is available, raises the appopriate `PropertyChanged` event to signal Solution Explorer that the UI needs to be updated.

The new and old types contain a lot of duplicated code. A later commit will refactor this code into shared base types.

Also, this change includes no support for the diagnostic node context menu (used to update the diagnostic severity in the rule set file) or targeting multiple frameworks. Those will come later as well.